smtp/requirements.txt
amqp==5.1.1
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
flower==2.0.0
flowers==0.1.3
humanize==4.7.0
idna==3.4
kombu==5.3.1
pdfminer3k==1.3.4
pika==1.3.2
ply==3.11
prometheus-client==0.17.0
prompt-toolkit==3.0.38
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
tornado==6.3.2
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
